+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Enh #20: Add `HasDecoding` trait and `decoding()` method to manage `decoding` attribute for HTML/SVG elements (@terabytesoftw)
 - Enh #21: Add `HasFetchpriority` trait and `fetchpriority()` method to manage `fetchpriority` attribute for HTML/SVG elements (@terabytesoftw)
 - Bug #22: Update documentation for `Crossorigin` and `ElementAttribute` enums to clarify attribute representation and compliance with MDN standards (@terabytesoftw)
+- Bug #23: Update attribute retrieval in tests to use `getAttribute()` method for consistency (@terabytesoftw)
 
 ## 0.4.0 December 27, 2025